Rename ProjectPortalPage.tsx, index.tsx
